chat bug solve

+++ app/src/main/java/com/example/e_mentoringrait/model/DataAttendence.java
@@ -2,7 +2,7 @@ package com.example.e_mentoringrait.model;
 
 public class DataAttendence {
     String SubjectName;
-    int Percentage;
+    float Percentage;
     int TotalClass;
     int PrasentClass;
     int AbsentClass;
@@ -10,7 +10,7 @@ public class DataAttendence {
     public DataAttendence() {
     }
 
-    public DataAttendence(String subjectName, int percentage, int totalClass, int prasentClass, int absentClass) {
+    public DataAttendence(String subjectName, float percentage, int totalClass, int prasentClass, int absentClass) {
         SubjectName = subjectName;
         Percentage = percentage;
         TotalClass = totalClass;
@@ -26,11 +26,11 @@ public class DataAttendence {
         SubjectName = subjectName;
     }
 
-    public int getPercentage() {
+    public float getPercentage() {
         return Percentage;
     }
 
-    public void setPercentage(int percentage) {
+    public void setPercentage(float percentage) {
         Percentage = percentage;
     }
